refactor: streamline editor key generation and deletion logic in activity and vote controllers
